LLM block constructor

Basic constructor for blocks that inherit from llm_block. Intended to be used for more specific block classes such as LLM-based transform or plot blocks.`)

new_llm_data_block(...)

# S3 method for class 'llm_data_block_proxy'
llm_block_server(x)

new_llm_flxtbl_block(...)

new_llm_gt_block(...)

new_llm_insights_block(...)

# S3 method for class 'llm_insights_block_proxy'
llm_block_ui(x)

# S3 method for class 'llm_insights_block_proxy'
llm_block_server(x)

new_llm_plot_block(...)

new_llm_transform_block(...)

new_llm_block(
  class,
  messages = list(),
  code = character(),
  ctor = sys.parent(),
  ...
)

llm_block_server(x)

# S3 method for class 'llm_block_proxy'
llm_block_server(x)

llm_block_ui(x)

# S3 method for class 'llm_block_proxy'
llm_block_ui(x)

Arguments

...

Forwarded to blockr.core::new_block()

x

An object inheriting from "block"

class

Block subclass

messages

See shinychat::chat_ui()

code

Code as generated by the LLM

ctor

String-valued constructor name or function/frame number (mostly for internal use or when defining constructors for virtual classes)

Value

All blocks constructed via new_llm_block() inherit from llm_block.